Adding labels
It's hard to get them drawing close to the camera. I'm thinking maybe we pick a random set of sats to label and periodically shuffle them. You would store a list of n indices and then every position update you move the active labels. Whenever a shuffle happens you switch the show for the previous labels and the new labels.